查询组织接口不过滤过期字段

+++ rosettaflow-mapper/src/main/java/com/moirae/rosettaflow/mapper/domain/UserMetaData.java
@@ -57,7 +57,7 @@ public class UserMetaData implements Serializable {
     /**
      * 授权值:按次数单位为（次）
      */
-    private Integer authValue;
+    private Long authValue;
     /**
      * 授权开始时间
      */
